requirements.txt
rsa
adb --no-binary :all:
cmd2
gnureadline
python-dateutil
PTable